+++ src/main/java/net/Gabou/projectatmosphere/mixin/client/DhSupportPipelineDiagnosticsMixin.java
@@ -53,7 +53,9 @@ public abstract class DhSupportPipelineDiagnosticsMixin {
         renderer.getCloudTarget().bindWrite(false);
         SimpleCloudsTornadoRenderer.INSTANCE.renderOpaque(
                 renderer, stack, projMat, partialTick, cloudColor[0], cloudColor[1], cloudColor[2],
-                renderer.getCloudTransparencyTarget().getDepthTextureId(), true
+                renderer.getCloudTransparencyTarget().getDepthTextureId(),
+                mc.getMainRenderTarget().getDepthTextureId(),
+                true
         );
 
         mc.getProfiler().pop();